Minesweeper Game in Python. Class Based
This file has the Minesweeper game I coded in Python using my first attempt at a Class within the language. The board is also a nested list: a list of rows. Where the format used to read a location is Rows[row][column]. In the Main.py file I named it board[row][column] to more easily understand what the variable is. This game does not have a UI, but is formatted to play on Command Prompt.