Added files via upload
This python program solves Sudoku puzzles using recursion. It guesses one possible value for a space, tests if it's valid, and tries the rest of the board with it. 
Main.py has the main functions of the code. Puzzles.py is a list of puzzles to test.
This code is executable in command prompt.